This PR alters the source for the ios FirebaseFirestore pod to instead use the pre-compiled .xcframework binaries provided by https://github.com/invertase/firestore-ios-sdk-frameworks

I set this up in my private repo fork months ago and do remember needing to update cocoapods and potentially bumping into one or two other minor changes I had to make, so this will need some testing and possibly adjustment to the Readme.

### Why

Currently the Firestore iOS SDK depends on some 500k lines of mostly C++, which when compiling as part of your Xcode build takes a long time - even more so in CI environments.

**Related Issues**

- [firebase/firebase-ios-sdk](https://github.com/firebase/firebase-ios-sdk)
  - [#4284](https://github.com/firebase/firebase-ios-sdk/issues/4284) `Adding FirebaseFirestore pod dependency adds minutes to build time`
- [FirebaseExtended/flutterfire](https://github.com/FirebaseExtended/flutterfire)
  - [#349](https://github.com/FirebaseExtended/flutterfire/issues/349) `[cloud_firestore] Xcode build extremely slow`
  
#### Before & After

Before and after timing below, timed when running Xcode build (with cache fully cleared) in a project with Firestore.

**Mac mini (2018) 6 cores**:

```
Before:    ~ 240s
After:     ~  45s
```

**GitHub Action CI 2 cores**:

```
Before:    ~ 551s
After:     ~ 174s
```